# openapi.yaml
info:
  description: RESTful API of Ingenium VenueServer that interfaces with test venues
  title: Ingenium VenueServer
  version: 3.0-non-ampcs
openapi: 3.1.0
paths:
  /api/v3/custom_script/halt:
    post:
      operationId: script_halt_api_v3_custom_script_halt_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptHaltBodyModel'
        required: true
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Halt custom script specified
      tags:
        - SCRIPT
  /api/v3/custom_script/start:
    post:
      description: >-
        This will start the custom script and return. Use status endpoint to get
        its status.
      operationId: script_start_api_v3_custom_script_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptStartBodyModel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptRunInfo'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Start execution of a custom script
      tags:
        - SCRIPT
  /api/v3/custom_script/status::
    get:
      operationId: script_status_api_v3_custom_script_status_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptStatusBodyModel'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptStatusResp'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Retrieve status information from currently executing script
      tags:
        - SCRIPT
  /api/v3/custom_script/{script_run_id}/files:
    get:
      operationId: script_file_api_v3_custom_script__script_run_id__files_get
      parameters:
        - in: path
          name: script_run_id
          required: true
          schema:
            title: Script Run Id
            type: string
      responses:
        '200':
          content:
            application/gzip: {}
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
      summary: Download the input, output, and log files of the custom script as tar.gz
      tags:
        - SCRIPT
  /api/v3/health:
    get:
      operationId: health_api_v3_health_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
      summary: Check health status of VenueServer
      tags:
        - HEALTH
tags:
  - description: Run custom scripts
    name: SCRIPT
  - description: Check service health
    name: HEALTH        
components:
  schemas:
    CustomScriptOutputs:
      properties:
        entries:
          allOf:
            - $ref: '#/components/schemas/ScriptEntriesModel'
          default: {}
          description: Custom script entries
          title: Entries
        output_array:
          default: []
          description: Custom script output array
          items:
            type: object
          title: Output Array
          type: array
        outputs:
          default: {}
          description: Custom script outputs fields
          title: Outputs
          type: object
      title: CustomScriptOutputs
      type: object
    CustomScriptStatus:
      description: An enumeration.
      enum:
        - PENDING
        - ERROR
        - PASS
        - FAIL
      title: CustomScriptStatus
      type: string
    ErrorResponse:
      properties:
        message:
          default: ''
          description: Error message
          title: Message
          type: string
      title: ErrorResponse
      type: object
    HealthStatus:
      properties:
        message:
          description: status message
          title: Message
          type: string
        status:
          allOf:
            - $ref: '#/components/schemas/HealthStatusEnum'
          description: health status
      required:
        - status
        - message
      title: HealthStatus
      type: object
    HealthStatusEnum:
      description: An enumeration.
      enum:
        - OK
        - ERROR
        - UNKNOWN
      title: HealthStatusEnum
      type: string
    ScriptEntriesModel:
      properties:
        entry_output_array:
          default: []
          description: Custom script entry output array
          items:
            type: object
          title: Entry Output Array
          type: array
        entry_outputs:
          default: {}
          description: Custom script entry outputs
          title: Entry Outputs
          type: object
        verification_status:
          allOf:
            - $ref: '#/components/schemas/VerificationStatus'
          default: PENDING
          description: Verification status for the specified entry
      title: ScriptEntriesModel
      type: object
    ScriptHaltBodyModel:
      properties:
        scriptRunId:
          description: custom script ID for the running custom script
          title: Scriptrunid
          type: string
      required:
        - scriptRunId
      title: ScriptHaltBodyModel
      type: object
    ScriptRunInfo:
      properties:
        scriptRunId:
          description: id of the running script
          title: Scriptrunid
          type: string
      required:
        - scriptRunId
      title: ScriptRunInfo
      type: object
    ScriptStartBodyModel:
      properties:
        inputs:
          default: {}
          description: Dictionary of custom script inputs and corresponding values
          title: Inputs
          type: object
        outputs:
          default: {}
          description: Dictionary of custom script outputs
          title: Outputs
          type: object
        scriptHash:
          description: SHA256 hash of custom script file
          title: Scripthash
          type: string
        scriptName:
          description: Name of custom script
          title: Scriptname
          type: string
        scriptPath:
          description: Relative path of custom script from custom script root directory
          title: Scriptpath
          type: string
      required:
        - scriptName
        - scriptPath
        - scriptHash
      title: ScriptStartBodyModel
      type: object
    ScriptStatusBodyModel:
      properties:
        scriptRunId:
          description: custom script ID for the running custom script
          title: Scriptrunid
          type: string
      required:
      - scriptRunId
      title: ScriptStatusBodyModel
      type: object      
    ScriptStatusResp:
      properties:
        custom_script_outputs:
          allOf:
            - $ref: '#/components/schemas/CustomScriptOutputs'
          default: {}
          description: >-
            Object that contains all current output values from the executed
            custom script
          title: Custom Script Outputs
        custom_script_status:
          allOf:
            - $ref: '#/components/schemas/CustomScriptStatus'
          default: PENDING
          description: The current status of the executed custom script
        logfile_lines:
          default: []
          description: last 25 lines of the custom script log file
          items:
            type: string
          title: Logfile Lines
          type: array
        logfile_path:
          default: ''
          description: The absolute path (on GDS host) to the custom script log file
          title: Logfile Path
          type: string
        logfile_url:
          default: ''
          description: >-
            Relative URL of download end point of logs.
            "custom_script/{script_run_id}/files"
          title: Logfile Url
          type: string
      title: ScriptStatusResp
      type: object
    VerificationStatus:
      description: An enumeration.
      enum:
        - PENDING
        - ERROR
        - PASS
        - FAIL
      title: VerificationStatus
      type: string          